requirements.txt
openai
pandas
numpy
tiktoken
psycopg2-binary
pgvector
python-dotenv